+++ app/Http/Controllers/Admin/AdminArticleController.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\File;
 use App\Http\Requests\CreateArticleRequest;
 use App\Models\Article;
 use App\Models\EnglishArticle;
+use App\Models\EnglishArticleCategory;
 use App\Models\FarsiArticle;
 use App\Models\Image;
 
@@ -19,22 +20,24 @@ class AdminArticleController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        $articles = Article::all();
-        return view("admin.articles.index", compact("articles"));
-    }
+    // public function index()
+    // {
+    //     $articles = Article::all();
+    //     return view("admin.articles.index", compact("articles"));
+    // }
 
     public function indexfarsi()
     {
-        $articles = Article::where("language", 0)->get();
-        return view("admin.articles.index", compact("articles"));
+        $articles = Article::all();
+        $lang = 0;
+        return view("admin.articles.index", compact("articles", "lang"));
     }
 
     public function indexenglish()
     {
-        $articles = Article::where("language", 1)->get();
-        return view("admin.articles.index", compact("articles"));
+        $articles = EnglishArticle::all();
+        $lang = 1;
+        return view("admin.articles.index", compact("articles", "lang"));
     }
 
     /**
@@ -42,10 +45,19 @@ class AdminArticleController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create($lang)
     {
-        $categories = ArticleCategory::all();
-        return view("admin.articles.create", compact("categories"));
+        // check if article is for farsi users
+        if ($lang == 0) {
+            $categories = ArticleCategory::all();
+            return view("admin.articles.create", compact("categories", "lang"));
+        }
+
+        // check if article is for english users
+        if ($lang == 1) {
+            $categories = EnglishArticleCategory::all();
+            return view("admin.articles.create", compact("categories", "lang"));
+        }
     }
 
     /**
@@ -56,29 +68,58 @@ class AdminArticleController extends Controller
      */
     public function store(CreateArticleRequest $request)
     {
-        $category = ArticleCategory::find($request->category_id)->first();
-        $article = new Article();
-        $article->title = $request->title;
-        $article->language = $request->language;
-        $article->category_id = $request->category_id;
-        $article->text = $request->text;
-        $article->meta_key_words = $request->meta_key_words;
-        $article->meta_descriptions = $request->meta_descriptions;
-        $article->save();
 
-        // saving image in image table
-        $image = new Image();
-        $imagename = time() . "." . $request->image->extension();
-        $filename = $article->title . "." . $category->id;
-        $request->image->move(public_path("photos/articles/$category->title/$filename/"), $imagename);
-        $image->name = $request->image_name;
-        $image->alt = $request->alt;
-        $image->uploader_id = auth()->user()->id;
-        $image->path = "photos/articles/$category->title/$filename/$imagename";
-        $article->images()->save($image);
-        // saving image in image table
+        // if article is for farsi users
+        if ($request->lang == 0) {
+            $category = ArticleCategory::find($request->category_id)->first();
+            $article = new Article();
+            $article->title = $request->title;
+            $article->category_id = $request->category_id;
+            $article->text = $request->text;
+            $article->meta_key_words = $request->meta_key_words;
+            $article->meta_descriptions = $request->meta_descriptions;
+            $article->save();
 
-        return redirect()->route("admin.articles.index")->with("success", "مقاله شما با موفقیت ساخته شد");
+            // saving image in image table
+            $image = new Image();
+            $imagename = time() . "." . $request->image->extension();
+            $filename = $article->title . "." . $category->id;
+            $request->image->move(public_path("photos/articles/$category->title/$filename/"), $imagename);
+            $image->name = $request->image_name;
+            $image->alt = $request->alt;
+            $image->uploader_id = auth()->user()->id;
+            $image->path = "photos/articles/$category->title/$filename/$imagename";
+            $article->images()->save($image);
+            // saving image in image table
+
+            return redirect()->route("admin.articles.farsi.index")->with("success", "مقاله شما با موفقیت ساخته شد");
+        }
+
+        // if article is for english users
+        if ($request->lang == 1) {
+            $category = EnglishArticleCategory::find($request->category_id)->first();
+            $article = new EnglishArticle();
+            $article->title = $request->title;
+            $article->category_id = $category->id;
+            $article->text = $request->text;
+            $article->meta_key_words = $request->meta_key_words;
+            $article->meta_descriptions = $request->meta_descriptions;
+            $article->save();
+
+            // saving image in image table
+            $image = new Image();
+            $imagename = time() . "." . $request->image->extension();
+            $filename = $article->title . "." . $category->id;
+            $request->image->move(public_path("photos/articles/$category->title/$filename/"), $imagename);
+            $image->name = $request->image_name;
+            $image->alt = $request->alt;
+            $image->uploader_id = auth()->user()->id;
+            $image->path = "photos/articles/$category->title/$filename/$imagename";
+            $article->images()->save($image);
+            // saving image in image table
+
+            return redirect()->route("admin.articles.english.index")->with("success", "مقاله شما با موفقیت ساخته شد");
+        }
     }
 
     /**
@@ -98,11 +139,19 @@ class AdminArticleController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit($id, $lang)
     {
-        $article = Article::find($id);
-        $categories = ArticleCategory::all();
-        return view("admin.articles.edit", compact("article", "categories"));
+        if ($lang == 0) {
+            $article = Article::find($id);
+            $categories = ArticleCategory::all();
+            return view("admin.articles.edit", compact("article", "categories", "lang"));
+        }
+
+        if ($lang == 1) {
+            $article = EnglishArticle::find($id);
+            $categories = EnglishArticleCategory::all();
+            return view("admin.articles.edit", compact("article", "categories", "lang"));
+        }
     }
 
 
@@ -115,14 +164,29 @@ class AdminArticleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $article =  Article::find($id);
-        $article->title = $request->title;
-        $article->category_id = $request->category_id;
-        $article->meta_key_words = $request->meta_key_words;
-        $article->meta_descriptions = $request->meta_descriptions;
-        $article->text = $request->text;
-        $article->save();
-        return redirect()->route("admin.articles.index")->with("success", "مقاله شما با موفقیت ویرایش شد");
+        // check if article is farsi
+        if ($request->lang == 0) {
+            $article =  Article::find($id);
+            $article->title = $request->title;
+            $article->category_id = $request->category_id;
+            $article->meta_key_words = $request->meta_key_words;
+            $article->meta_descriptions = $request->meta_descriptions;
+            $article->text = $request->text;
+            $article->save();
+            return redirect()->route("admin.articles.farsi.index")->with("success", "مقاله شما با موفقیت ویرایش شد");
+        }
+
+        // check if article is english
+        if ($request->lang == 1) {
+            $article =  EnglishArticle::find($id);
+            $article->title = $request->title;
+            $article->category_id = $request->category_id;
+            $article->meta_key_words = $request->meta_key_words;
+            $article->meta_descriptions = $request->meta_descriptions;
+            $article->text = $request->text;
+            $article->save();
+            return redirect()->route("admin.articles.english.index")->with("success", "مقاله شما با موفقیت ویرایش شد");
+        }
     }
 
     /**
@@ -131,14 +195,28 @@ class AdminArticleController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $article = Article::find($id);
-        File::delete($article->images[0]->path);
-        $path = pathinfo($article->images[0]->path)["dirname"];
-        rmdir($path);
-        $article->images()->delete();
-        $article->delete();
-        return redirect()->back()->with("success", "مقاله شما با موفقیت حذف شد");
+        // check if article is farsi
+        if ($request->lang == 0) {
+            $article = Article::find($id);
+            File::delete($article->images[0]->path);
+            $path = pathinfo($article->images[0]->path)["dirname"];
+            rmdir($path);
+            $article->images()->delete();
+            $article->delete();
+            return redirect()->back()->with("success", "مقاله شما با موفقیت حذف شد");
+        }
+
+        // check if article is english
+        if ($request->lang == 1) {
+            $article = EnglishArticle::find($id);
+            File::delete($article->images[0]->path);
+            $path = pathinfo($article->images[0]->path)["dirname"];
+            rmdir($path);
+            $article->images()->delete();
+            $article->delete();
+            return redirect()->back()->with("success", "مقاله شما با موفقیت حذف شد");
+        }
     }
 }
